webapp/requirements.txt
streamlit
torch
torchvision
Pillow
numpy
matplotlib
opencv-python